Data dictionary of Brazil's census data

Open on a browser the data dictionary of Brazil's census data.

data_dictionary(year = 2010, dataset = NULL, showProgress = TRUE, cache = TRUE)

Arguments

year

Numeric. Year of reference in the format yyyy. Defaults to 2010.

dataset

Character. The dataset of data dictionary to be opened. Options include c("population", "households", "families", "mortality", "emigration", "tracts").

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

cache

Logical. Whether the function should read the data cached locally, which is much faster. Defaults to TRUE. The first time the user runs the function, censobr will download the file and store it locally so that the file only needs to be download once. If FALSE, the function will download the data again and overwrite the local file.

Value

Returns NULL and opens .html or .pdf file on the browser

See also

Examples

# Open data dictionary on browser
data_dictionary(year = 2010, dataset = 'population', showProgress = FALSE)
#> Downloading data and storing it locally for future use.
#> NULL

data_dictionary(year = 1980, dataset = 'households', showProgress = FALSE)
#> Downloading data and storing it locally for future use.
#> NULL

data_dictionary(year = 2010, dataset = 'tracts', showProgress = FALSE)
#> Downloading data and storing it locally for future use.
#> NULL