Download microdata of death records from Brazil's census

Download microdata of death records from Brazil's census. Data collected in the sample component of the questionnaire.

read_deaths(
  year = 2010,
  columns = NULL,
  as_data_frame = FALSE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

year

Numeric. Year of reference in the format yyyymm. Defaults to 2010.

columns

String. A vector of column names to keep. The rest of the columns are not read. Defaults to NULL and read all columns.

as_data_frame

Logical. Whether the function should return a data.frame or an Arrow Table (Default).

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

cache

Logical. Whether the function should read the data cached locally in future calls, which is much faster. Defaults to TRUE. If FALSE, the function will download the file from the internet.

Value

An ArrowObject Dataset or a "data.frame" object.

See also

Examples

# return data as arrow table
df <- read_deaths(year = 2010)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  16%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |================================                                      |  45%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  52%  |                                                                              |================================================                      |  68%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%
head(df)
#> Table
#> 6 rows x 26 columns
#> $code_muni <string>
#> $code_state <string>
#> $abbrev_state <string>
#> $name_state <string>
#> $code_region <string>
#> $name_region <string>
#> $code_weighting <string>
#> $V0001 <string>
#> $V0002 <string>
#> $V0011 <string>
#> $V0300 <double>
#> $V0010 <double>
#> $V1001 <string>
#> $V1002 <string>
#> $V1003 <string>
#> $V1004 <string>
#> $V1006 <string>
#> $V0703 <string>
#> $V0704 <string>
#> $V7051 <double>
#> $V7052 <double>
#> $M0703 <string>
#> $M0704 <string>
#> $M7051 <string>
#> $M7052 <string>
#> $V1005 <string>
#> 
#> See $metadata for additional Schema metadata

# return data as data.frame
df <- read_deaths(year = 2010, as_data_frame = TRUE)
head(df)
#>   code_muni code_state abbrev_state name_state code_region name_region
#> 1   1100015         11           RO   Rondônia           1       Norte
#> 2   1100015         11           RO   Rondônia           1       Norte
#> 3   1100015         11           RO   Rondônia           1       Norte
#> 4   1100015         11           RO   Rondônia           1       Norte
#> 5   1100015         11           RO   Rondônia           1       Norte
#> 6   1100015         11           RO   Rondônia           1       Norte
#>   code_weighting V0001 V0002         V0011   V0300     V0010 V1001 V1002 V1003
#> 1  1100015001001    11 00015 1100015001001 1061597 10.811192     1    02   006
#> 2  1100015001001    11 00015 1100015001001 1460419 10.418260     1    02   006
#> 3  1100015001001    11 00015 1100015001001 2903970 13.104927     1    02   006
#> 4  1100015001001    11 00015 1100015001001 3508873  9.694958     1    02   006
#> 5  1100015001001    11 00015 1100015001001 3922312  5.334208     1    02   006
#> 6  1100015001001    11 00015 1100015001001 4219372 10.002353     1    02   006
#>   V1004 V1006 V0703 V0704 V7051 V7052 M0703 M0704 M7051 M7052 V1005
#> 1    00     1    08     2    34    NA     2     2     2     2     1
#> 2    00     1    09     2     2    NA     2     2     2     2     1
#> 3    00     2    07     1    77    NA     2     2     2     2     8
#> 4    00     2    02     2    66    NA     2     2     2     2     8
#> 5    00     1    07     1    NA     2     2     2     2     2     1
#> 6    00     1    08     2    79    NA     2     2     2     2     1