Download microdata of family records from Brazil's census

Download microdata of family records from Brazil's census. Data collected in the sample component of the questionnaire.

read_families(
  year = 2000,
  columns = NULL,
  add_labels = NULL,
  as_data_frame = FALSE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

year

Numeric. Year of reference in the format yyyymm. Defaults to 2000.

columns

String. A vector of column names to keep. The rest of the columns are not read. Defaults to NULL and read all columns.

add_labels

Character. Whether the function should add labels to the responses of categorical variables. When add_labels = "pt", the function adds labels in Portuguese. Defaults to NULL.

as_data_frame

Logical. When FALSE (Default), the function returns an Arrow Dataset, which allows users to work with larger-than-memory data. If TRUE, the function returns data.frame.

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

cache

Logical. Whether the function should read the data cached locally, which is much faster. Defaults to TRUE. The first time the user runs the function, censobr will download the file and store it locally so that the file only needs to be download once. If FALSE, the function will download the data again and overwrite the local file.

Value

An arrow Dataset or a "data.frame" object.

See also

Examples

# return data as arrow Dataset
df <- read_families(year = 2000,
                    showProgress = FALSE)
#> Downloading data. File will be stored locally at: /home/runner/.cache/R/censobr/data_release_v0.1.0

head(df) |> dplyr::collect()
#>   code_muni code_state abbrev_state name_state code_region name_region
#> 1   1100015         11           RO   Rondônia           1       Norte
#> 2   1100015         11           RO   Rondônia           1       Norte
#> 3   1100015         11           RO   Rondônia           1       Norte
#> 4   1100015         11           RO   Rondônia           1       Norte
#> 5   1100015         11           RO   Rondônia           1       Norte
#> 6   1100015         11           RO   Rondônia           1       Norte
#>   code_weighting V0102 V1002 V1003   V0103     V0104       V0105   V0300 V0404
#> 1  1100015001001    11  1102 11006 1100015 110001505 11000150599 5188550     1
#> 2  1100015001001    11  1102 11006 1100015 110001505 11000150599 1252632     1
#> 3  1100015001001    11  1102 11006 1100015 110001505 11000150599 3472851     1
#> 4  1100015001001    11  1102 11006 1100015 110001505 11000150599 2030293     1
#> 5  1100015001001    11  1102 11006 1100015 110001505 11000150599 4065906     1
#> 6  1100015001001    11  1102 11006 1100015 110001505 11000150599 1932380     1
#>   V1001 V1004         AREAP CODV0404 CODV0404_2 V4614B CODV4615B V4614C
#> 1     1  <NA> 1100015001001        0         11   3086        10   2950
#> 2     1  <NA> 1100015001001        0         01    301        04    301
#> 3     1  <NA> 1100015001001        0         03    250        04    250
#> 4     1  <NA> 1100015001001        0         02    740        06    740
#> 5     1  <NA> 1100015001001        0         02    953        07    953
#> 6     1  <NA> 1100015001001        0         02    300        04    300
#>   CODV4615C V4616_7400 CODV4615_7400 V7400 CODV7400 V7400A CODV7400A V7400B
#> 1        12     514.33            06     6       06      3        03      3
#> 2        07     150.50            03     2       02      1        01      1
#> 3        07      83.33            03     3       03      2        02      1
#> 4        09     246.67            04     3       03      1        01      2
#> 5        10     238.25            04     4       04      1        01      3
#> 6        07     100.00            03     3       03      1        01      2
#>   CODV7400B      P001
#> 1        03  8.382025
#> 2        01  9.491570
#> 3        01  8.938103
#> 4        02 12.508102
#> 5        03 11.325142
#> 6        02  9.425035