Download microdata of population records from Brazil's census

Download microdata of population records from Brazil's census. Data collected in the sample component of the questionnaire.

read_population(
  year = 2010,
  columns = NULL,
  as_data_frame = FALSE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

year

Numeric. Year of reference in the format yyyymm. Defaults to 2010.

columns

String. A vector of column names to keep. The rest of the columns are not read. Defaults to NULL and read all columns.

as_data_frame

Logical. Whether the function should return a data.frame or an Arrow Table (Default).

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

cache

Logical. The first time the user runs the function, censobr will download the file from the internet and store it locally so that the file only needs to be download once. When the cache parameter is set to TRUE (Default), the function will read the cached data, which is much faster. If FALSE, the function will download the file from the internet again and overwrite an eventual local file. The data is stored locally at tools::R_user_dir('censobr_v0.1', 'cache').

Value

An ArrowObject Dataset or a "data.frame" object.

See also

Examples

# return data as arrow table
df <- read_population(year = 2010)
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |==                                                                    |   4%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |=========                                                             |  14%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |================                                                      |  24%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |=======================                                               |  34%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |==============================                                        |  44%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |============================================                          |  64%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |===================================================                   |  74%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |==========================================================            |  84%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  91%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |=================================================================     |  94%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%
head(df)
#> Table
#> 6 rows x 251 columns
#> $code_muni <string>
#> $code_state <string>
#> $abbrev_state <string>
#> $name_state <string>
#> $code_region <string>
#> $name_region <string>
#> $code_weighting <string>
#> $V0001 <string>
#> $V0002 <string>
#> $V0011 <string>
#> $V0300 <double>
#> $V0010 <double>
#> $V1001 <string>
#> $V1002 <string>
#> $V1003 <string>
#> $V1004 <string>
#> $V1006 <string>
#> $V0502 <string>
#> $V0504 <string>
#> $V0601 <string>
#> $V6033 <double>
#> $V6036 <double>
#> $V6037 <double>
#> $V6040 <string>
#> $V0606 <string>
#> $V0613 <string>
#> $V0614 <string>
#> $V0615 <string>
#> $V0616 <string>
#> $V0617 <string>
#> $V0618 <string>
#> $V0619 <string>
#> $V0620 <string>
#> $V0621 <string>
#> $V0622 <string>
#> $V6222 <string>
#> $V6224 <string>
#> $V0623 <string>
#> $V0624 <string>
#> $V0625 <string>
#> $V6252 <string>
#> $V6254 <string>
#> $V6256 <string>
#> $V0626 <string>
#> $V6262 <string>
#> $V6264 <string>
#> $V6266 <string>
#> $V0627 <string>
#> $V0628 <string>
#> $V0629 <string>
#> $V0630 <string>
#> $V0631 <string>
#> $V0632 <string>
#> $V0633 <string>
#> $V0634 <string>
#> $V0635 <string>
#> $V6400 <string>
#> $V6352 <string>
#> $V6354 <string>
#> $V6356 <string>
#> $V0636 <string>
#> $V6362 <string>
#> $V6364 <string>
#> $V6366 <string>
#> $V0637 <string>
#> $V0638 <string>
#> $V0639 <string>
#> $V0640 <string>
#> $V0641 <string>
#> $V0642 <string>
#> $V0643 <string>
#> $V0644 <string>
#> $V0645 <string>
#> $V6461 <string>
#> $V6471 <string>
#> $V0648 <string>
#> $V0649 <string>
#> $V0650 <string>
#> $V0651 <string>
#> $V6511 <double>
#> $V6513 <double>
#> $V6514 <double>
#> $V0652 <string>
#> $V6521 <double>
#> $V6524 <double>
#> $V6525 <double>
#> $V6526 <double>
#> $V6527 <double>
#> $V6528 <double>
#> $V6529 <double>
#> $V6530 <double>
#> $V6531 <double>
#> $V6532 <double>
#> $V0653 <double>
#> $V0654 <string>
#> $V0655 <string>
#> $V0656 <string>
#> $V0657 <string>
#> $V0658 <string>
#> $V0659 <string>
#> $V6591 <double>
#> $V0660 <string>
#> $V6602 <string>
#> $V6604 <string>
#> $V6606 <string>
#> $V0661 <string>
#> $V0662 <string>
#> $V0663 <string>
#> $V6631 <double>
#> $V6632 <double>
#> $V6633 <double>
#> $V0664 <string>
#> $V6641 <double>
#> $V6642 <double>
#> $V6643 <double>
#> $V0665 <string>
#> $V6660 <double>
#> $V6664 <string>
#> $V0667 <string>
#> $V0668 <string>
#> $V6681 <string>
#> $V6682 <string>
#> $V0669 <string>
#> $V6691 <double>
#> $V6692 <double>
#> $V6693 <double>
#> $V6800 <double>
#> $V0670 <string>
#> $V0671 <string>
#> $V6900 <string>
#> $V6910 <string>
#> $V6920 <string>
#> $V6930 <string>
#> $V6940 <string>
#> $V6121 <string>
#> $V0604 <string>
#> $V0605 <string>
#> $V5020 <string>
#> $V5060 <double>
#> $V5070 <double>
#> $V5080 <double>
#> $V6462 <string>
#> $V6472 <string>
#> $V5110 <string>
#> $V5120 <string>
#> $V5030 <string>
#> $V5040 <string>
#> $V5090 <string>
#> $V5100 <string>
#> $V5130 <string>
#> $M0502 <string>
#> $M0601 <string>
#> $M6033 <string>
#> $M0606 <string>
#> $M0613 <string>
#> $M0614 <string>
#> $M0615 <string>
#> $M0616 <string>
#> $M0617 <string>
#> $M0618 <string>
#> $M0619 <string>
#> $M0620 <string>
#> $M0621 <string>
#> $M0622 <string>
#> $M6222 <string>
#> $M6224 <string>
#> $M0623 <string>
#> $M0624 <string>
#> $M0625 <string>
#> $M6252 <string>
#> $M6254 <string>
#> $M6256 <string>
#> $M0626 <string>
#> $M6262 <string>
#> $M6264 <string>
#> $M6266 <string>
#> $M0627 <string>
#> $M0628 <string>
#> $M0629 <string>
#> $M0630 <string>
#> $M0631 <string>
#> $M0632 <string>
#> $M0633 <string>
#> $M0634 <string>
#> $M0635 <string>
#> $M6352 <string>
#> $M6354 <string>
#> $M6356 <string>
#> $M0636 <string>
#> $M6362 <string>
#> $M6364 <string>
#> $M6366 <string>
#> $M0637 <string>
#> $M0638 <string>
#> $M0639 <string>
#> $M0640 <string>
#> $M0641 <string>
#> $M0642 <string>
#> $M0643 <string>
#> $M0644 <string>
#> $M0645 <string>
#> $M6461 <string>
#> $M6471 <string>
#> $M0648 <string>
#> $M0649 <string>
#> $M0650 <string>
#> $M0651 <string>
#> $M6511 <string>
#> $M0652 <string>
#> $M6521 <string>
#> $M0653 <string>
#> $M0654 <string>
#> $M0655 <string>
#> $M0656 <string>
#> $M0657 <string>
#> $M0658 <string>
#> $M0659 <string>
#> $M6591 <string>
#> $M0660 <string>
#> $M6602 <string>
#> $M6604 <string>
#> $M6606 <string>
#> $M0661 <string>
#> $M0662 <string>
#> $M0663 <string>
#> $M6631 <string>
#> $M6632 <string>
#> $M6633 <string>
#> $M0664 <string>
#> $M6641 <string>
#> $M6642 <string>
#> $M6643 <string>
#> $M0665 <string>
#> $M6660 <string>
#> $M0667 <string>
#> $M0668 <string>
#> $M6681 <string>
#> $M6682 <string>
#> $M0669 <string>
#> $M6691 <string>
#> $M6692 <string>
#> $M6693 <string>
#> $M0670 <string>
#> $M0671 <string>
#> $M6800 <string>
#> $M6121 <string>
#> $M0604 <string>
#> $M0605 <string>
#> $M6462 <string>
#> $M6472 <string>
#> $V1005 <string>
#> 
#> See $metadata for additional Schema metadata