Download microdata of household records from Brazil's census

Download microdata of household records from Brazil's census. Data collected in the sample component of the questionnaire.

read_households(
  year = 2010,
  columns = NULL,
  add_labels = NULL,
  as_data_frame = FALSE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

year

Numeric. Year of reference in the format yyyymm. Defaults to 2010.

columns

String. A vector of column names to keep. The rest of the columns are not read. Defaults to NULL and read all columns.

add_labels

Character. Whether the function should add labels to the responses of categorical variables. When add_labels = "pt", the function adds labels in Portuguese. Defaults to NULL.

as_data_frame

Logical. When FALSE (Default), the function returns an Arrow Dataset, which allows users to work with larger-than-memory data. If TRUE, the function returns data.frame.

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

cache

Logical. Whether the function should read the data cached locally, which is much faster. Defaults to TRUE. The first time the user runs the function, censobr will download the file and store it locally so that the file only needs to be download once. If FALSE, the function will download the data again and overwrite the local file.

Value

An arrow Dataset or a "data.frame" object.

See also

Examples

# return data as arrow Dataset
df <- read_households(year = 2010,
                      showProgress = FALSE)
#> Downloading data. File will be stored locally at: /home/runner/.cache/R/censobr/data_release_v0.1.0

dplyr::glimpse(df)
#> FileSystemDataset with 1 Parquet file
#> 6,192,332 rows x 83 columns
#> $ code_muni      <string> "1100015", "1100015", "1100015", "1100015", "1100015",…
#> $ code_state     <string> "11", "11", "11", "11", "11", "11", "11", "11", "11", …
#> $ abbrev_state   <string> "RO", "RO", "RO", "RO", "RO", "RO", "RO", "RO", "RO", …
#> $ name_state     <string> "Rondônia", "Rondônia", "Rondônia", "Rondônia", "Rondô…
#> $ code_region    <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1",…
#> $ name_region    <string> "Norte", "Norte", "Norte", "Norte", "Norte", "Norte", …
#> $ code_weighting <string> "1100015001001", "1100015001001", "1100015001001", "11…
#> $ V0001          <string> "11", "11", "11", "11", "11", "11", "11", "11", "11", …
#> $ V0002          <string> "00015", "00015", "00015", "00015", "00015", "00015", …
#> $ V0011          <string> "1100015001001", "1100015001001", "1100015001001", "11…
#> $ V0300          <double> 3624, 6647, 14596, 15524, 15840, 29337, 38146, 45797, …
#> $ V0010          <double> 8.705865, 9.818689, 9.495608, 9.122642, 11.999527, 12.…
#> $ V1001          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1",…
#> $ V1002          <string> "02", "02", "02", "02", "02", "02", "02", "02", "02", …
#> $ V1003          <string> "006", "006", "006", "006", "006", "006", "006", "006"…
#> $ V1004          <string> "00", "00", "00", "00", "00", "00", "00", "00", "00", …
#> $ V1006          <string> "1", "1", "1", "2", "1", "1", "1", "1", "1", "2", "1",…
#> $ V4001          <string> "01", "01", "01", "01", "01", "01", "01", "01", "01", …
#> $ V4002          <string> "11", "11", "11", "11", "11", "11", "11", "11", "11", …
#> $ V0201          <string> "3", "1", "1", "1", "1", "1", "2", "1", "1", "4", NA, …
#> $ V2011          <double> 60, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V2012          <double> 0.11765, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ V0202          <string> "4", "1", "3", "3", "1", "3", "1", "6", "3", "1", NA, …
#> $ V0203          <double> 4, 7, 7, 10, 4, 6, 9, 6, 6, 7, NA, 3, 3, 5, 8, 10, 5, …
#> $ V6203          <double> 0.5, 0.1, 1.4, 0.3, 0.8, 0.8, 0.3, 0.3, 0.5, 0.4, NA, …
#> $ V0204          <double> 1, 1, 3, 2, 2, 2, 3, 1, 2, 2, NA, 2, 2, 4, 2, 2, 1, 3,…
#> $ V6204          <double> 2.0, 1.0, 3.3, 1.5, 1.5, 2.5, 1.0, 2.0, 1.5, 1.5, NA, …
#> $ V0205          <string> "1", "2", "1", "1", "1", "1", "2", "1", "1", "1", NA, …
#> $ V0206          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "2…
#> $ V0207          <string> "3", "3", "3", "3", "3", "3", "3", "4", "3", "3", NA, …
#> $ V0208          <string> "01", "01", "02", "02", "02", "02", "01", "02", "02", …
#> $ V0209          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", NA, …
#> $ V0210          <string> "1", "1", "1", "3", "1", "1", "1", "1", "1", "3", NA, …
#> $ V0211          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", NA, …
#> $ V0212          <string> "2", "1", "1", "2", "1", "1", "1", "1", "1", "1", NA, …
#> $ V0213          <string> "2", "1", "1", "1", "2", "1", "1", "1", "1", "1", NA, …
#> $ V0214          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", NA, …
#> $ V0215          <string> "2", "2", "2", "2", "2", "2", "1", "2", "1", "2", NA, …
#> $ V0216          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", NA, …
#> $ V0217          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", NA, …
#> $ V0218          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", NA, …
#> $ V0219          <string> "2", "1", "1", "1", "2", "2", "1", "2", "2", "2", NA, …
#> $ V0220          <string> NA, "1", "1", "1", NA, NA, "1", NA, NA, NA, NA, NA, NA…
#> $ V0221          <string> "1", "1", "1", "1", "1", "2", "1", "1", "1", "2", NA, …
#> $ V0222          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "1", NA, …
#> $ V0301          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", NA, …
#> $ V0401          <double> 2, 1, 10, 3, 3, 5, 3, 2, 3, 3, 1, 3, 2, 5, 2, 3, 1, 3,…
#> $ V0402          <string> "1", "1", "2", "1", "2", "2", "1", "2", "2", "2", NA, …
#> $ V0701          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", NA, …
#> $ V6529          <double> 1000, 1800, 3689, 1550, 2510, 970, 1530, 1020, 1900, 1…
#> $ V6530          <double> 1.96078, 3.52941, 7.23333, 3.03922, 4.92157, 1.90196, …
#> $ V6531          <double> 500.00, 1800.00, 368.90, 516.67, 836.67, 194.00, 510.0…
#> $ V6532          <double> 0.98039, 3.52941, 0.72333, 1.01307, 1.64052, 0.38039, …
#> $ V6600          <string> "2", "1", "3", "2", "3", "2", "2", "2", "2", "2", NA, …
#> $ V6210          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", NA, …
#> $ M0201          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M2011          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0202          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0203          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0204          <string> "2", "1", "2", "1", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0205          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0206          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0207          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0208          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0209          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0210          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0211          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0212          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0213          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0214          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0215          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0216          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0217          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0218          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0219          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0220          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0221          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0222          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0301          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0401          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0402          <string> "2", "1", "2", "2", "2", "2", "2", "2", "2", "2", "1",…
#> $ M0701          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ V1005          <string> "1", "1", "1", "8", "1", "1", "1", "1", "1", "8", "1",…
#> Call `print()` for full schema details