Download microdata of population records from Brazil's census

Download microdata of population records from Brazil's census.

Usage

read_population(
  year = 2010,
  columns = NULL,
  as_data_frame = FALSE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

year

Numeric. Year of reference in the format yyyymm. Defaults to 2010.

columns

String. A vector of column names to keep. The rest of the columns are not read. Defaults to NULL and read all columns.

as_data_frame

Logical. Whether the function should return a data.frame or an Arrow Table (Default).

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

cache

Logical. Whether the function should read the data cached locally in future calls, which is much faster. Defaults to TRUE. If FALSE, the function will download the file from the internet.

Value

An Arrow table or a "data.frame" object.

See also

Examples

if (FALSE)  if (interactive()) {
# return data as arrow table
df <- read_population(year = 2010)
head(df)

# # return data as data.frame
# df <- read_population(year = 2010, as_data_frame = TRUE)
# head(df)
}