Download microdata of population records from Brazil's census

Download microdata of population records from Brazil's census. Data collected in the sample component of the questionnaire.

read_population(
  year = 2010,
  columns = NULL,
  add_labels = NULL,
  as_data_frame = FALSE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

year

Numeric. Year of reference in the format yyyymm. Defaults to 2010.

columns

String. A vector of column names to keep. The rest of the columns are not read. Defaults to NULL and read all columns.

add_labels

Character. Whether the function should add labels to the responses of categorical variables. When add_labels = "pt", the function adds labels in Portuguese. Defaults to NULL.

as_data_frame

Logical. When FALSE (Default), the function returns an Arrow Dataset, which allows users to work with larger-than-memory data. If TRUE, the function returns data.frame.

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

cache

Logical. Whether the function should read the data cached locally, which is much faster. Defaults to TRUE. The first time the user runs the function, censobr will download the file and store it locally so that the file only needs to be download once. If FALSE, the function will download the data again and overwrite the local file.

Value

An arrow Dataset or a "data.frame" object.

See also

Examples

# return data as arrow Dataset
df <- read_population(year = 2010,
                      showProgress = FALSE)
#> Downloading data. File will be stored locally at: /home/runner/.cache/R/censobr/data_release_v0.1.0

dplyr::glimpse(df)
#> FileSystemDataset with 1 Parquet file
#> 20,635,472 rows x 251 columns
#> $ code_muni      <string> "1100015", "1100015", "1100015", "1100015", "1100015",…
#> $ code_state     <string> "11", "11", "11", "11", "11", "11", "11", "11", "11", …
#> $ abbrev_state   <string> "RO", "RO", "RO", "RO", "RO", "RO", "RO", "RO", "RO", …
#> $ name_state     <string> "Rondônia", "Rondônia", "Rondônia", "Rondônia", "Rondô…
#> $ code_region    <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1",…
#> $ name_region    <string> "Norte", "Norte", "Norte", "Norte", "Norte", "Norte", …
#> $ code_weighting <string> "1100015001001", "1100015001001", "1100015001001", "11…
#> $ V0001          <string> "11", "11", "11", "11", "11", "11", "11", "11", "11", …
#> $ V0002          <string> "00015", "00015", "00015", "00015", "00015", "00015", …
#> $ V0011          <string> "1100015001001", "1100015001001", "1100015001001", "11…
#> $ V0300          <double> 3624, 3624, 6647, 14596, 14596, 14596, 14596, 14596, 1…
#> $ V0010          <double> 8.705865, 8.705865, 9.818689, 9.495608, 9.495608, 9.49…
#> $ V1001          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1",…
#> $ V1002          <string> "02", "02", "02", "02", "02", "02", "02", "02", "02", …
#> $ V1003          <string> "006", "006", "006", "006", "006", "006", "006", "006"…
#> $ V1004          <string> "00", "00", "00", "00", "00", "00", "00", "00", "00", …
#> $ V1006          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1",…
#> $ V0502          <string> "01", "02", "01", "02", "04", "04", "04", "04", "04", …
#> $ V0504          <string> "01", "02", "01", "02", "09", "08", "07", "06", "05", …
#> $ V0601          <string> "1", "2", "1", "2", "1", "1", "2", "2", "2", "1", "1",…
#> $ V6033          <double> 22, 17, 38, 44, 11, 16, 17, 22, 26, 54, 11, 28, 27, 15…
#> $ V6036          <double> 22, 17, 38, 44, 11, 16, 17, 22, 26, 54, 11, 28, 27, 15…
#> $ V6037          <double> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6040          <string> "1", "1", "1", "2", "1", "1", "1", "1", "1", "1", "1",…
#> $ V0606          <string> "4", "4", "1", "1", "4", "4", "4", "4", "4", "4", "4",…
#> $ V0613          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0614          <string> "4", "4", "4", "4", "4", "4", "4", "4", "4", "3", "4",…
#> $ V0615          <string> "4", "4", "4", "4", "4", "4", "4", "4", "4", "4", "4",…
#> $ V0616          <string> "4", "4", "4", "4", "4", "4", "4", "4", "4", "4", "4",…
#> $ V0617          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ V0618          <string> "3", "1", "3", "3", "1", "1", "1", "1", "1", "3", "1",…
#> $ V0619          <string> "1", NA, "3", "3", NA, NA, NA, NA, NA, "3", NA, "3", "…
#> $ V0620          <string> NA, NA, "1", "1", NA, NA, NA, NA, NA, "1", NA, "1", "1…
#> $ V0621          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0622          <string> NA, NA, "1", "1", NA, NA, NA, NA, NA, "1", NA, "1", "1…
#> $ V6222          <string> NA, NA, "4100000", "5100000", NA, NA, NA, NA, NA, "210…
#> $ V6224          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0623          <string> NA, NA, "034", "018", NA, NA, NA, NA, NA, "018", NA, "…
#> $ V0624          <string> "022", NA, "025", "004", NA, NA, NA, NA, NA, "004", NA…
#> $ V0625          <string> NA, NA, NA, "1", NA, NA, NA, NA, NA, "1", NA, "1", "1"…
#> $ V6252          <string> NA, NA, NA, "1100000", NA, NA, NA, NA, NA, "1100000", …
#> $ V6254          <string> NA, NA, NA, "1101500", NA, NA, NA, NA, NA, "1101500", …
#> $ V6256          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0626          <string> NA, NA, NA, "1", NA, NA, NA, NA, NA, "1", NA, "1", "1"…
#> $ V6262          <string> NA, NA, NA, "1100000", NA, NA, NA, NA, NA, "1100000", …
#> $ V6264          <string> NA, NA, NA, "1101500", NA, NA, NA, NA, NA, "1101500", …
#> $ V6266          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0627          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "2", "1",…
#> $ V0628          <string> "3", "1", "3", "3", "1", "1", "1", "3", "3", "4", "1",…
#> $ V0629          <string> NA, "05", NA, NA, "05", "05", "07", NA, NA, NA, "05", …
#> $ V0630          <string> NA, "09", NA, NA, "06", "08", NA, NA, NA, NA, "05", NA…
#> $ V0631          <string> NA, NA, NA, NA, NA, NA, "1", NA, NA, NA, NA, NA, NA, "…
#> $ V0632          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0633          <string> "07", NA, "07", "05", NA, NA, NA, "05", "10", NA, NA, …
#> $ V0634          <string> "2", NA, "1", NA, NA, NA, NA, NA, "1", NA, NA, NA, "1"…
#> $ V0635          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6400          <string> "1", "1", "2", "1", "1", "1", "2", "1", "3", "1", "1",…
#> $ V6352          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6354          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6356          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0636          <string> NA, "1", NA, NA, "1", "1", "1", NA, NA, NA, "1", NA, N…
#> $ V6362          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6364          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6366          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0637          <string> "1", "1", "2", "1", "3", "3", "3", "3", "3", "1", "3",…
#> $ V0638          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0639          <string> "4", "4", NA, "4", NA, NA, NA, NA, NA, "4", NA, NA, NA…
#> $ V0640          <string> "5", "5", "5", "5", "5", "5", "5", "5", "5", "1", "5",…
#> $ V0641          <string> "1", "2", "1", "1", "2", "2", "2", "2", "1", "1", "2",…
#> $ V0642          <string> NA, "2", NA, NA, "2", "2", "2", "2", NA, NA, "2", NA, …
#> $ V0643          <string> NA, "2", NA, NA, "2", "2", "2", "2", NA, NA, "2", NA, …
#> $ V0644          <string> NA, "2", NA, NA, "2", "2", "2", "2", NA, NA, "2", NA, …
#> $ V0645          <string> "1", NA, "1", "1", NA, NA, NA, NA, "1", "1", NA, "1", …
#> $ V6461          <string> "7522", NA, "0000", "9111", NA, NA, NA, NA, "3344", "5…
#> $ V6471          <string> "31000", NA, "00000", "97000", NA, NA, NA, NA, "86002"…
#> $ V0648          <string> "1", NA, "5", "4", NA, NA, NA, NA, "1", "1", NA, "1", …
#> $ V0649          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0650          <string> NA, NA, "3", "3", NA, NA, NA, NA, NA, NA, NA, NA, "3",…
#> $ V0651          <string> "1", NA, "1", "1", NA, NA, NA, NA, "1", "1", NA, "1", …
#> $ V6511          <double> 1000, NA, 1800, 250, NA, NA, NA, NA, 650, 700, NA, 900…
#> $ V6513          <double> 1000, NA, 1800, 250, NA, NA, NA, NA, 650, 700, NA, 900…
#> $ V6514          <double> 1.96, NA, 3.53, 0.49, NA, NA, NA, NA, 1.27, 1.37, NA, …
#> $ V0652          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6521          <double> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6524          <double> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6525          <double> 1000, NA, 1800, 250, NA, NA, NA, NA, 650, 700, NA, 900…
#> $ V6526          <double> 1.96078, NA, 3.52941, 0.49020, NA, NA, NA, NA, 1.27451…
#> $ V6527          <double> 1000, 0, 1800, 419, 0, 0, 0, 510, 650, 700, 0, 900, 51…
#> $ V6528          <double> 1.96078, 0.00000, 3.52941, 0.82157, 0.00000, 0.00000, …
#> $ V6529          <double> 1000, 1000, 1800, 3689, 3689, 3689, 3689, 3689, 3689, …
#> $ V6530          <double> 1.96078, 1.96078, 3.52941, 7.23333, 7.23333, 7.23333, …
#> $ V6531          <double> 500.00, 500.00, 1800.00, 368.90, 368.90, 368.90, 368.9…
#> $ V6532          <double> 0.98039, 0.98039, 3.52941, 0.72333, 0.72333, 0.72333, …
#> $ V0653          <double> 50, NA, 60, 30, NA, NA, NA, NA, 40, 40, NA, 40, 25, NA…
#> $ V0654          <string> NA, "2", NA, NA, "2", "2", "2", "2", NA, NA, "2", NA, …
#> $ V0655          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0656          <string> "0", "0", "0", "0", "0", "0", "0", "1", "0", "0", "0",…
#> $ V0657          <string> "0", "0", "0", "1", "0", "0", "0", "0", "0", "0", "0",…
#> $ V0658          <string> "0", "0", "0", "0", "0", "0", "0", "0", "0", "0", "0",…
#> $ V0659          <string> "0", "0", "0", "0", "0", "0", "0", "0", "0", "0", "0",…
#> $ V6591          <double> 0, 0, 0, 169, 0, 0, 0, 510, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ V0660          <string> "2", NA, "1", "2", NA, NA, NA, NA, "2", "2", NA, "2", …
#> $ V6602          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6604          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6606          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0661          <string> "1", NA, NA, "1", NA, NA, NA, NA, "1", "1", NA, "1", "…
#> $ V0662          <string> "2", NA, NA, "3", NA, NA, NA, NA, "3", "2", NA, "2", "…
#> $ V0663          <string> NA, "2", NA, "1", NA, NA, "2", "2", "2", NA, NA, NA, "…
#> $ V6631          <double> NA, NA, NA, 3, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ V6632          <double> NA, NA, NA, 5, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ V6633          <double> NA, 0, NA, 8, NA, NA, 0, 0, 0, NA, NA, NA, 0, NA, 2, N…
#> $ V0664          <string> NA, NA, NA, "1", NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ V6641          <double> NA, NA, NA, 3, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ V6642          <double> NA, NA, NA, 5, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ V6643          <double> NA, NA, NA, 8, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ V0665          <string> NA, NA, NA, "1", NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ V6660          <double> NA, NA, NA, 11, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6664          <string> NA, "0", NA, "0", NA, NA, "0", "0", "0", NA, NA, NA, "…
#> $ V0667          <string> NA, NA, NA, "1", NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ V0668          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6681          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6682          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V0669          <string> NA, "2", NA, "2", NA, NA, "2", "2", "2", NA, NA, NA, "…
#> $ V6691          <double> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6692          <double> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V6693          <double> NA, 0, NA, 0, NA, NA, 0, 0, 0, NA, NA, NA, 0, NA, 1, N…
#> $ V6800          <double> NA, 0, NA, 8, NA, NA, 0, 0, 0, NA, NA, NA, 0, NA, 3, N…
#> $ V0670          <string> "2", "1", "1", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ V0671          <string> "02", NA, NA, "04", "04", "04", "04", "04", "04", "04"…
#> $ V6900          <string> "1", "2", "1", "1", "2", "2", "2", "2", "1", "1", "2",…
#> $ V6910          <string> "1", NA, "1", "1", NA, NA, NA, NA, "1", "1", NA, "1", …
#> $ V6920          <string> "1", "2", "1", "1", "2", "2", "2", "2", "1", "1", "2",…
#> $ V6930          <string> "1", NA, "4", "3", NA, NA, NA, NA, "1", "1", NA, "1", …
#> $ V6940          <string> "3", NA, NA, "2", NA, NA, NA, NA, "3", "3", NA, "1", "…
#> $ V6121          <string> "110", "110", "490", "110", "110", "110", "110", "110"…
#> $ V0604          <string> "2", "2", "3", "2", "1", "1", "1", "1", "1", "2", "2",…
#> $ V0605          <string> NA, NA, NA, NA, "02", "02", "02", "02", "02", NA, NA, …
#> $ V5020          <string> "01", "01", "01", "01", "01", "01", "01", "01", "01", …
#> $ V5060          <double> 2, 2, 1, 10, 10, 10, 10, 10, 10, 10, 10, 10, 10, 3, 3,…
#> $ V5070          <double> 500.00, 500.00, 1800.00, 368.90, 368.90, 368.90, 368.9…
#> $ V5080          <double> 0.98039, 0.98039, 3.52941, 0.72333, 0.72333, 0.72333, …
#> $ V6462          <string> "7711", NA, "0000", "5121", NA, NA, NA, NA, "4221", "5…
#> $ V6472          <string> "36010", NA, "00000", "95000", NA, NA, NA, NA, "85012"…
#> $ V5110          <string> "1", NA, "2", "2", NA, NA, NA, NA, "1", "1", NA, "1", …
#> $ V5120          <string> "1", NA, "2", "2", NA, NA, NA, NA, "1", "1", NA, "1", …
#> $ V5030          <string> "3", "3", "1", "3", "3", "3", "3", "3", "3", "3", "3",…
#> $ V5040          <string> "1", "1", "0", "1", "1", "1", "1", "1", "1", "1", "1",…
#> $ V5090          <string> "1", "1", NA, "4", "4", "4", "4", "4", "4", "4", "4", …
#> $ V5100          <string> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ V5130          <string> "01", "02", "01", "02", "09", "08", "07", "06", "05", …
#> $ M0502          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0601          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6033          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0606          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0613          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0614          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0615          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0616          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0617          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0618          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0619          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0620          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0621          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0622          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6222          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6224          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0623          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0624          <string> "1", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0625          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6252          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6254          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6256          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0626          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6262          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6264          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6266          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0627          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0628          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0629          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0630          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0631          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0632          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0633          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0634          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0635          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6352          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6354          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6356          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0636          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6362          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6364          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6366          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0637          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0638          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0639          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0640          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0641          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0642          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0643          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0644          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0645          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6461          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6471          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0648          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0649          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0650          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0651          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6511          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0652          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6521          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0653          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0654          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0655          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0656          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0657          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0658          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0659          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6591          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0660          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6602          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6604          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6606          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0661          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0662          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0663          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6631          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6632          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6633          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0664          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6641          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6642          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6643          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0665          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6660          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0667          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0668          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6681          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6682          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0669          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6691          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6692          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6693          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0670          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0671          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6800          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6121          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0604          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M0605          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6462          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ M6472          <string> "2", "2", "2", "2", "2", "2", "2", "2", "2", "2", "2",…
#> $ V1005          <string> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1",…
#> Call `print()` for full schema details