Download microdata of population records from Brazil's census

Download microdata of population records from Brazil's census. Data collected in the sample component of the questionnaire.

read_population(
  year = 2010,
  columns = NULL,
  add_labels = NULL,
  as_data_frame = FALSE,
  showProgress = TRUE,
  cache = TRUE
)

Arguments

year

Numeric. Year of reference in the format yyyymm. Defaults to 2010.

columns

String. A vector of column names to keep. The rest of the columns are not read. Defaults to NULL and read all columns.

add_labels

Character. Whether the function should add labels to the responses of categorical variables. When add_labels = "pt", the function adds labels in Portuguese. Defaults to NULL.

as_data_frame

Logical. When FALSE (Default), the function returns an Arrow Dataset, which allows users to work with larger-than-memory data. If TRUE, the function returns data.frame.

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

cache

Logical. Whether the function should read the data cached locally, which is much faster. Defaults to TRUE. The first time the user runs the function, censobr will download the file and store it locally so that the file only needs to be download once. If FALSE, the function will download the data again and overwrite the local file.

Value

An arrow Dataset or a "data.frame" object.

See also

Examples

# return data as arrow Dataset
df <- read_population(year = 2010,
                      showProgress = FALSE)
#> Downloading data. File will be stored locally at: /home/runner/.cache/R/censobr/data_release_v0.1.0

head(df) |> dplyr::collect()
#>   code_muni code_state abbrev_state name_state code_region name_region
#> 1   1100015         11           RO   Rondônia           1       Norte
#> 2   1100015         11           RO   Rondônia           1       Norte
#> 3   1100015         11           RO   Rondônia           1       Norte
#> 4   1100015         11           RO   Rondônia           1       Norte
#> 5   1100015         11           RO   Rondônia           1       Norte
#> 6   1100015         11           RO   Rondônia           1       Norte
#>   code_weighting V0001 V0002         V0011 V0300    V0010 V1001 V1002 V1003
#> 1  1100015001001    11 00015 1100015001001  3624 8.705865     1    02   006
#> 2  1100015001001    11 00015 1100015001001  3624 8.705865     1    02   006
#> 3  1100015001001    11 00015 1100015001001  6647 9.818689     1    02   006
#> 4  1100015001001    11 00015 1100015001001 14596 9.495608     1    02   006
#> 5  1100015001001    11 00015 1100015001001 14596 9.495608     1    02   006
#> 6  1100015001001    11 00015 1100015001001 14596 9.495608     1    02   006
#>   V1004 V1006 V0502 V0504 V0601 V6033 V6036 V6037 V6040 V0606 V0613 V0614 V0615
#> 1    00     1    01    01     1    22    22    NA     1     4  <NA>     4     4
#> 2    00     1    02    02     2    17    17    NA     1     4  <NA>     4     4
#> 3    00     1    01    01     1    38    38    NA     1     1  <NA>     4     4
#> 4    00     1    02    02     2    44    44    NA     2     1  <NA>     4     4
#> 5    00     1    04    09     1    11    11    NA     1     4  <NA>     4     4
#> 6    00     1    04    08     1    16    16    NA     1     4  <NA>     4     4
#>   V0616 V0617 V0618 V0619 V0620 V0621 V0622   V6222 V6224 V0623 V0624 V0625
#> 1     4     2     3     1  <NA>  <NA>  <NA>    <NA>  <NA>  <NA>   022  <NA>
#> 2     4     2     1  <NA>  <NA>  <NA>  <NA>    <NA>  <NA>  <NA>  <NA>  <NA>
#> 3     4     2     3     3     1  <NA>     1 4100000  <NA>   034   025  <NA>
#> 4     4     2     3     3     1  <NA>     1 5100000  <NA>   018   004     1
#> 5     4     2     1  <NA>  <NA>  <NA>  <NA>    <NA>  <NA>  <NA>  <NA>  <NA>
#> 6     4     2     1  <NA>  <NA>  <NA>  <NA>    <NA>  <NA>  <NA>  <NA>  <NA>
#>     V6252   V6254 V6256 V0626   V6262   V6264 V6266 V0627 V0628 V0629 V0630
#> 1    <NA>    <NA>  <NA>  <NA>    <NA>    <NA>  <NA>     1     3  <NA>  <NA>
#> 2    <NA>    <NA>  <NA>  <NA>    <NA>    <NA>  <NA>     1     1    05    09
#> 3    <NA>    <NA>  <NA>  <NA>    <NA>    <NA>  <NA>     1     3  <NA>  <NA>
#> 4 1100000 1101500  <NA>     1 1100000 1101500  <NA>     1     3  <NA>  <NA>
#> 5    <NA>    <NA>  <NA>  <NA>    <NA>    <NA>  <NA>     1     1    05    06
#> 6    <NA>    <NA>  <NA>  <NA>    <NA>    <NA>  <NA>     1     1    05    08
#>   V0631 V0632 V0633 V0634 V0635 V6400 V6352 V6354 V6356 V0636 V6362 V6364 V6366
#> 1  <NA>  <NA>    07     2  <NA>     1  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>
#> 2  <NA>  <NA>  <NA>  <NA>  <NA>     1  <NA>  <NA>  <NA>     1  <NA>  <NA>  <NA>
#> 3  <NA>  <NA>    07     1  <NA>     2  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>
#> 4  <NA>  <NA>    05  <NA>  <NA>     1  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>
#> 5  <NA>  <NA>  <NA>  <NA>  <NA>     1  <NA>  <NA>  <NA>     1  <NA>  <NA>  <NA>
#> 6  <NA>  <NA>  <NA>  <NA>  <NA>     1  <NA>  <NA>  <NA>     1  <NA>  <NA>  <NA>
#>   V0637 V0638 V0639 V0640 V0641 V0642 V0643 V0644 V0645 V6461 V6471 V0648 V0649
#> 1     1  <NA>     4     5     1  <NA>  <NA>  <NA>     1  7522 31000     1  <NA>
#> 2     1  <NA>     4     5     2     2     2     2  <NA>  <NA>  <NA>  <NA>  <NA>
#> 3     2  <NA>  <NA>     5     1  <NA>  <NA>  <NA>     1  0000 00000     5  <NA>
#> 4     1  <NA>     4     5     1  <NA>  <NA>  <NA>     1  9111 97000     4  <NA>
#> 5     3  <NA>  <NA>     5     2     2     2     2  <NA>  <NA>  <NA>  <NA>  <NA>
#> 6     3  <NA>  <NA>     5     2     2     2     2  <NA>  <NA>  <NA>  <NA>  <NA>
#>   V0650 V0651 V6511 V6513 V6514 V0652 V6521 V6524 V6525   V6526 V6527   V6528
#> 1  <NA>     1  1000  1000  1.96  <NA>    NA    NA  1000 1.96078  1000 1.96078
#> 2  <NA>  <NA>    NA    NA    NA  <NA>    NA    NA    NA      NA     0 0.00000
#> 3     3     1  1800  1800  3.53  <NA>    NA    NA  1800 3.52941  1800 3.52941
#> 4     3     1   250   250  0.49  <NA>    NA    NA   250 0.49020   419 0.82157
#> 5  <NA>  <NA>    NA    NA    NA  <NA>    NA    NA    NA      NA     0 0.00000
#> 6  <NA>  <NA>    NA    NA    NA  <NA>    NA    NA    NA      NA     0 0.00000
#>   V6529   V6530  V6531   V6532 V0653 V0654 V0655 V0656 V0657 V0658 V0659 V6591
#> 1  1000 1.96078  500.0 0.98039    50  <NA>  <NA>     0     0     0     0     0
#> 2  1000 1.96078  500.0 0.98039    NA     2  <NA>     0     0     0     0     0
#> 3  1800 3.52941 1800.0 3.52941    60  <NA>  <NA>     0     0     0     0     0
#> 4  3689 7.23333  368.9 0.72333    30  <NA>  <NA>     0     1     0     0   169
#> 5  3689 7.23333  368.9 0.72333    NA     2  <NA>     0     0     0     0     0
#> 6  3689 7.23333  368.9 0.72333    NA     2  <NA>     0     0     0     0     0
#>   V0660 V6602 V6604 V6606 V0661 V0662 V0663 V6631 V6632 V6633 V0664 V6641 V6642
#> 1     2  <NA>  <NA>  <NA>     1     2  <NA>    NA    NA    NA  <NA>    NA    NA
#> 2  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>     2    NA    NA     0  <NA>    NA    NA
#> 3     1  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>    NA    NA    NA  <NA>    NA    NA
#> 4     2  <NA>  <NA>  <NA>     1     3     1     3     5     8     1     3     5
#> 5  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>    NA    NA    NA  <NA>    NA    NA
#> 6  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>    NA    NA    NA  <NA>    NA    NA
#>   V6643 V0665 V6660 V6664 V0667 V0668 V6681 V6682 V0669 V6691 V6692 V6693 V6800
#> 1    NA  <NA>    NA  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>    NA    NA    NA    NA
#> 2    NA  <NA>    NA     0  <NA>  <NA>  <NA>  <NA>     2    NA    NA     0     0
#> 3    NA  <NA>    NA  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>    NA    NA    NA    NA
#> 4     8     1    11     0     1  <NA>  <NA>  <NA>     2    NA    NA     0     8
#> 5    NA  <NA>    NA  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>    NA    NA    NA    NA
#> 6    NA  <NA>    NA  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>    NA    NA    NA    NA
#>   V0670 V0671 V6900 V6910 V6920 V6930 V6940 V6121 V0604 V0605 V5020 V5060
#> 1     2    02     1     1     1     1     3   110     2  <NA>    01     2
#> 2     1  <NA>     2  <NA>     2  <NA>  <NA>   110     2  <NA>    01     2
#> 3     1  <NA>     1     1     1     4  <NA>   490     3  <NA>    01     1
#> 4     2    04     1     1     1     3     2   110     2  <NA>    01    10
#> 5     2    04     2  <NA>     2  <NA>  <NA>   110     1    02    01    10
#> 6     2    04     2  <NA>     2  <NA>  <NA>   110     1    02    01    10
#>    V5070   V5080 V6462 V6472 V5110 V5120 V5030 V5040 V5090 V5100 V5130 M0502
#> 1  500.0 0.98039  7711 36010     1     1     3     1     1  <NA>    01     2
#> 2  500.0 0.98039  <NA>  <NA>  <NA>  <NA>     3     1     1  <NA>    02     2
#> 3 1800.0 3.52941  0000 00000     2     2     1     0  <NA>  <NA>    01     2
#> 4  368.9 0.72333  5121 95000     2     2     3     1     4  <NA>    02     2
#> 5  368.9 0.72333  <NA>  <NA>  <NA>  <NA>     3     1     4  <NA>    09     2
#> 6  368.9 0.72333  <NA>  <NA>  <NA>  <NA>     3     1     4  <NA>    08     2
#>   M0601 M6033 M0606 M0613 M0614 M0615 M0616 M0617 M0618 M0619 M0620 M0621 M0622
#> 1     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 2     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 3     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 4     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 5     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 6     2     2     2     2     2     2     2     2     2     2     2     2     2
#>   M6222 M6224 M0623 M0624 M0625 M6252 M6254 M6256 M0626 M6262 M6264 M6266 M0627
#> 1     2     2     2     1     2     2     2     2     2     2     2     2     2
#> 2     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 3     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 4     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 5     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 6     2     2     2     2     2     2     2     2     2     2     2     2     2
#>   M0628 M0629 M0630 M0631 M0632 M0633 M0634 M0635 M6352 M6354 M6356 M0636 M6362
#> 1     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 2     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 3     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 4     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 5     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 6     2     2     2     2     2     2     2     2     2     2     2     2     2
#>   M6364 M6366 M0637 M0638 M0639 M0640 M0641 M0642 M0643 M0644 M0645 M6461 M6471
#> 1     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 2     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 3     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 4     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 5     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 6     2     2     2     2     2     2     2     2     2     2     2     2     2
#>   M0648 M0649 M0650 M0651 M6511 M0652 M6521 M0653 M0654 M0655 M0656 M0657 M0658
#> 1     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 2     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 3     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 4     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 5     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 6     2     2     2     2     2     2     2     2     2     2     2     2     2
#>   M0659 M6591 M0660 M6602 M6604 M6606 M0661 M0662 M0663 M6631 M6632 M6633 M0664
#> 1     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 2     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 3     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 4     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 5     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 6     2     2     2     2     2     2     2     2     2     2     2     2     2
#>   M6641 M6642 M6643 M0665 M6660 M0667 M0668 M6681 M6682 M0669 M6691 M6692 M6693
#> 1     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 2     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 3     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 4     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 5     2     2     2     2     2     2     2     2     2     2     2     2     2
#> 6     2     2     2     2     2     2     2     2     2     2     2     2     2
#>   M0670 M0671 M6800 M6121 M0604 M0605 M6462 M6472 V1005
#> 1     2     2     2     2     2     2     2     2     1
#> 2     2     2     2     2     2     2     2     2     1
#> 3     2     2     2     2     2     2     2     2     1
#> 4     2     2     2     2     2     2     2     2     1
#> 5     2     2     2     2     2     2     2     2     1
#> 6     2     2     2     2     2     2     2     2     1